Don't document that __capability expands to __attribute__((cheri_capability))
The attribute should be considered an implementation, with __capability
being the proper way to annotate capabilities. I personally believe the
attribute should just go and __capability be made a keyword, which would
also allow it to be used in more interesting ways (such as a new
operator, "& __capability" or "__capability &", to get a capability
pointer to a given variable in a robust way that doesn't rely on casts
being directly on the address-of expression). MSVC would also never use
the __attribute__ syntax as that's a GNU extension, so would either
treat it as a keyword or use a __declspec equivalent.